Remove unused rendering functions.
